+++ src/Greenleaf/PageAction.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @package Greenleaf
+ * @license http://opensource.org/licenses/MIT
+ */
+
+declare(strict_types=1);
+
+namespace DecodeLabs\Greenleaf;
+
+interface PageAction extends Action
+{
+    public int $priority { get; }
+}