+++ src/Greenleaf/Router.php
@@ -31,10 +31,10 @@ interface Router
     /**
      * Find route for leaf URI
      *
-     * @param array<string, string|Stringable|int|float|null> $params
+     * @param array<string, string|Stringable|int|float|null> $parameters
      */
     public function matchOut(
         string|LeafUrl $uri,
-        ?array $params = null
+        ?array $parameters = null
     ): ?Hit;
 }